Make complete method for legacy code and move to separate file
